+++ RELEASE_NOTES.md
@@ -3,6 +3,15 @@ The Narrative Interface allows users to craft KBase Narratives using a combinati
 
 This is built on the IPython Notebook (more notes will follow).
 
+### Release 9/25/2014
+__Changes__
+- Fixed a bug that prevented a few Narrative functions from working if the
+user's workspace was empty.
+
+### Release 9/24/2014
+__Changes__
+- Updated workspace URL from the Narrative Interface
+
 ### Release 8/14/2014
 __Changes__
 - Added functionality to Communities services